fix: prerender Article and FAQ JSON-LD into answer-page HTML
The answers JSON-LD only existed via react-helmet after client-side
hydration, so raw-HTML fetchers — ChatGPT-User and PerplexityBot, the
exact channel these pages target — never saw the Article or the new
FAQPage markup. The builders move to a pure module (answersJsonLd.ts)
shared by the page component and the prerender script, and
emitAnswersPages now injects both script blocks into the emitted head.
